+++ frontend/src/lib/components/RecipePopularity/DislikeButton.svelte
