+++ frontend/src/lib/components/RecipePopularity/LikeButton.svelte
